Refactored trigger command and moved its functionality into the sdk package
